refactor(tests): rename test files by subject and drop duplicated cases
phase2 -> changelog_version, phase3 -> tag_publish, new_and_release ->
release_pr; member_discovery merged into configless; doctor and publish
tests moved next to their peers; the two invalid-fragment tests merged
into one; seven cli.rs --json tests removed because the json_contract
snapshots already pin every assertion. Tests: 413 before, 405 after.
